add quoteId to orders, add logs
