+++ crates/model/src/lib.rs
@@ -4,6 +4,7 @@ pub mod app_id;
 pub mod auction;
 pub mod bytes_hex;
 pub mod order;
+pub mod quote;
 pub mod ratio_as_decimal;
 pub mod signature;
 pub mod solver_competition;
